consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["simon.islam <simon.islam@brilliant.com.bd>"]

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.2.1"
mysql-connector-python = "^8.0.29"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
